RF+TST: move make_tempfile to utils and add tests
remove non-essential variable assignment from knows_annex test
