+++ datalad/tests/test_utils.py
@@ -34,6 +34,7 @@ from ..utils import file_basename
 from ..utils import expandpath, is_explicit_path
 from ..utils import knows_annex
 from ..utils import any_re_search
+from ..utils import make_tempfile
 from ..support.annexrepo import AnnexRepo
 
 from nose.tools import ok_, eq_, assert_false, assert_equal, assert_true
@@ -375,9 +376,16 @@ def test_is_explicit_path():
 def test_knows_annex(here, there):
     from datalad.support.gitrepo import GitRepo
     from datalad.support.annexrepo import AnnexRepo
-    git = GitRepo(path=here, create=True)
+    GitRepo(path=here, create=True)
     assert_false(knows_annex(here))
-    annex = AnnexRepo(path=here, create=True)
+    AnnexRepo(path=here, create=True)
     assert_true(knows_annex(here))
-    gitclone = GitRepo(path=there, url=here, create=True)
+    GitRepo(path=there, url=here, create=True)
     assert_true(knows_annex(there))
+
+
+def test_make_tempfile():
+    # check if mkdir, content conflict caught
+    with assert_raises(ValueError):
+        with make_tempfile(content="blah", mkdir=True):
+            pass
